Add files via upload
I have now added a login and signup section of the app that saves the data locally, I have used multithreading to speed up encryption of larger files but it is still fairly slow (40 seconds for 1gb). If trying this code change the number of threads to suit your system.